Move `mutating_list_iterate` to framework lib
So that the utility is accessible in base libraries
