+++ lib/framework/object_list_iteration.h
@@ -1,0 +1,56 @@
+/*
+	This file is part of Warzone 2100.
+	Copyright (C) 2024  Warzone 2100 Project
+
+	Warzone 2100 is free software; you can redistribute it and/or modify
+	it under the terms of the GNU General Public License as published by
+	the Free Software Foundation; either version 2 of the License, or
+	(at your option) any later version.
+
+	Warzone 2100 is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+	GNU General Public License for more details.
+
+	You should have received a copy of the GNU General Public License
+	along with Warzone 2100; if not, write to the Free Software
+	Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
+*/
+/** @file
+ *  Basic routines for list iteration of various in-game objects (resources, sounds, droids, etc.).
+ */
+#pragma once
+
+#include <list>
+#include <type_traits>
+#include <iterator>
+
+enum class IterationResult
+{
+	BREAK_ITERATION,
+	CONTINUE_ITERATION
+};
+
+// Common iteration helper for lists of game objects
+// with an ability to execute loop body handlers which can
+// possibly invalidate the any iterator in the range `[begin(), currentIterator]`.
+template <typename ObjectType, typename MaybeErasingLoopBodyHandler>
+void mutating_list_iterate(std::list<ObjectType*>& list, MaybeErasingLoopBodyHandler handler)
+{
+	static_assert(std::is_same<std::result_of_t<MaybeErasingLoopBodyHandler(ObjectType*)>, IterationResult>::value,
+		"Loop body handler should return IterationResult");
+
+	typename std::remove_reference_t<decltype(list)>::iterator it = list.begin(), itNext;
+	while (it != list.end())
+	{
+		itNext = std::next(it);
+		// Can possibly invalidate `it` and anything before it.
+		const auto res = handler(*it);
+		if (res == IterationResult::BREAK_ITERATION)
+		{
+			break;
+		}
+		it = itNext;
+	}
+}
+
